Refactor primitives
