+++ src/python/model_api/visualizer/primitive/__init__.py
@@ -1,0 +1,10 @@
+"""Primitive classes."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .bounding_box import BoundingBox
+from .overlay import Overlay
+from .primitive import Primitive
+
+__all__ = ["Primitive", "BoundingBox", "Overlay"]
