+++ src/python/model_api/visualizer/primitive/bounding_box.py
@@ -1,23 +1,13 @@
-"""Base class for primitives."""
+"""Bounding box primitive."""
 
-# Copyright (C) 2024 Intel Corporation
+# Copyright (C) 2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-
-import numpy as np
-import PIL
 from PIL import Image, ImageDraw
 
-
-class Primitive(ABC):
-    """Primitive class."""
-
-    @abstractmethod
-    def compute(self, image: Image) -> Image:
-        pass
+from .primitive import Primitive
 
 
 class BoundingBox(Primitive):
@@ -71,27 +61,3 @@ class BoundingBox(Primitive):
             draw.text((0, 0), self.label, fill="white")
             image.paste(label_image, (self.x1, self.y1))
         return image
-
-
-class Overlay(Primitive):
-    """Overlay primitive.
-
-    Useful for XAI and Anomaly Maps.
-
-    Args:
-        image (PIL.Image | np.ndarray): Image to be overlaid.
-        opacity (float): Opacity of the overlay.
-    """
-
-    def __init__(self, image: PIL.Image | np.ndarray, opacity: float = 0.4) -> None:
-        self.image = self._to_pil(image)
-        self.opacity = opacity
-
-    def _to_pil(self, image: PIL.Image | np.ndarray) -> PIL.Image:
-        if isinstance(image, np.ndarray):
-            return PIL.Image.fromarray(image)
-        return image
-
-    def compute(self, image: PIL.Image) -> PIL.Image:
-        _image = self.image.resize(image.size)
-        return PIL.Image.blend(image, _image, self.opacity)
